g_packages/official_py_docker/docker/requirements.txt
# pip==18.1
# ansicolors==1.1.8
# Cython
# Keras==2.1.6
# Theano==1.0.1
# anndata==0.5.9
# bokeh==0.12.15
# xlrd==1.1.0
# colour==0.1.5
# h5py==2.7.1
# hdbscan==0.8.13
# ipython==6.2.1
# joblib==0.11
# # matplotlib==2.1.2
# natsort==5.2.0
# networkx==1.11
# numba==0.37.0
# numpy==1.14.0
# pandas==0.22.0
# scikit-learn==0.19.1
# scipy==1.0.0
# seaborn==0.8.1
# statsmodels==0.8.0
# tables==3.4.2
# python-igraph==0.7.1.post6
# louvain==0.6.1
# ipdb==0.11
# tqdm==4.28.1
# sortedcontainers==2.1.0
# umap-learn==0.3.8

pip
ansicolors
Cython
Keras
Theano
anndata
bokeh
xlrd
colour
h5py
hdbscan
ipython
joblib
natsort
networkx
numba
numpy
pandas
scikit-learn
scipy
seaborn
statsmodels
tables
python-igraph
louvain
ipdb
tqdm
sortedcontainers
umap-learn
